# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Team360 Health Check Application REST API. Assesses team health across 11 dimensions with hierarchical org support.",
        "title": "Team360 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/managers/{managerId}/teams/action-items": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns each supervised team's open (non-done) action item count. Requires manager-or-above role; caller may only request their own summary.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "action-items"
                ],
                "summary": "Get open action item counts across a manager's teams",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Manager ID",
                        "name": "managerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.TeamsActionSummaryResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — can only view your own summary",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch/read action summaries",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/teams/{teamId}/action-items": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns action items for a team, optionally filtered by status and/or assessment period. Requires team membership.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "action-items"
                ],
                "summary": "List a team's action items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (open, in_progress, done)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by assessment period",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ActionItemsResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch/scan/read action items",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new action item for a team. If assignedTo is set, the user must be a member of the team. Requires team membership.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "action-items"
                ],
                "summary": "Create a team action item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Action item",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.CreateActionItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "id, status, createdAt",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request, invalid dueDate, or assignedTo not a team member",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to validate assignee or create action item",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/teams/{teamId}/action-items/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an action item belonging to a team. Requires team membership.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "action-items"
                ],
                "summary": "Delete a team action item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Action item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "404": {
                        "description": "Action item not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete action item",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially updates an action item's fields. If assignedTo is set, the user must be a member of the team. Requires team membership.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "action-items"
                ],
                "summary": "Update a team action item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Action item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.UpdateActionItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request, invalid status/dueDate, or assignedTo not a team member",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Action item not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to validate assignee or update action item",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.ActionItemResponse": {
            "type": "object",
            "properties": {
                "assessmentPeriod": {
                    "type": "string"
                },
                "assignedTo": {
                    "type": "string"
                },
                "assigneeName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "createdByName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dimensionId": {
                    "type": "string"
                },
                "dimensionName": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "teamId": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.ActionItemsResponse": {
            "type": "object",
            "properties": {
                "actionItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ActionItemResponse"
                    }
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.CreateActionItemRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "assessmentPeriod": {
                    "type": "string"
                },
                "assignedTo": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dimensionId": {
                    "type": "string"
                },
                "dueDate": {
                    "description": "ISO date string YYYY-MM-DD",
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.TeamActionSummaryResponse": {
            "type": "object",
            "properties": {
                "openCount": {
                    "type": "integer"
                },
                "teamId": {
                    "type": "string"
                },
                "teamName": {
                    "type": "string"
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.TeamsActionSummaryResponse": {
            "type": "object",
            "properties": {
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.TeamActionSummaryResponse"
                    }
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.UpdateActionItemRequest": {
            "type": "object",
            "properties": {
                "assessmentPeriod": {
                    "type": "string"
                },
                "assignedTo": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dimensionId": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}